Renamed text-utils to strings for clarity
