+++ src/components/pokemon-card.tsx
@@ -1,4 +1,4 @@
-import { capitaliseFirstLetter } from '@/utils/text-utils';
+import { capitaliseFirstLetter } from '@/utils/strings';
 import Image from 'next/image';
 import { Pokemon } from '@/lib/types/types';
 import { ASSET_PATHS } from '@/lib/constants';
